% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_new_ops.R
\name{-.gg}
\alias{-.gg}
\title{Pipe ggplot objects using `-`}
\usage{
\method{-}{gg}(lhs, rhs)
}
\description{
This is a pure alias for `%>%` but it has a different operator precedence
which makes it convenient when you want to pipe a ggplot chain into a function
without using parenthesis around the call.
}
