% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_pprint.R
\name{pprint}
\alias{pprint}
\title{Pipe friendly printing}
\usage{
pprint(.data, .fun, ...)
}
\arguments{
\item{.data}{An object}

\item{.fun}{An expression or a function to apply on \code{.data} so the output can be printed,
supports formula notation through \code{purrr::as_mapper}.}

\item{...}{Additional parameters passed to .fun}
}
\description{
\code{pprint} makes it easy to print information about the pipe chain's status.
}
\examples{
iris \%>\%
pprint(~"hello")           \%>\%
pprint("hi")               \%>\% # simple expressions work as well
head(2)                    \%>\%
transform(Species = NULL)  \%>\%
pprint(rowSums,na.rm=TRUE) \%>\%
pprint(~dplyr::rename_all(.[1:2],toupper)) \%>\%
pprint(dim)
}
