% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/07_rm_pipe.R
\name{rm_pipe}
\alias{rm_pipe}
\title{Remove custom pipe}
\usage{
rm_pipe(x)
}
\arguments{
\item{x}{pipe to remove, as a raw name}
}
\description{
used to remove pipes created with \code{add_pipe}, it's not allowed to remove
a predefined pipe
}
