% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_add_pipe.R
\name{add_pipe}
\alias{add_pipe}
\title{build a new pipe in one command}
\usage{
add_pipe(new_pipe, new_body)
}
\arguments{
\item{new_pipe}{name}

\item{new_body}{call, the variable body used in this call holds the initial
call}
}
\value{
the operator is created and *magrittr*'s namespace is modified to
  make the new operator compatible. NULL is returned.
}
\description{
build a new pipe in one command
}
\examples{
# if we wanted to recreate existing operators
add_pipe(`\%T2>\%`, call("{", body, quote(.)))
iris \%T2>\% {message("side effect")} \%>\% head(2)
add_pipe(`\%W2>\%`, substitute(
    {options(warn = -1); on.exit(options(warn = w)); b},
    list(w = options()$warn, b = body)))
data.frame(a = c(1,-1)) \%W2>\% transform(a = sqrt(a))
add_pipe(`\%P2>\%`, substitute({. <- print(b);cat("\\n");.}, list(b = body)))
iris \%P>\% head(3) \%>\% head(2)
}
